requirements.txt
tensorflow>=2.10
scikit-learn
pillow
numpy
torch
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
